fix(opsgenie): allow title to be used as a Message
this also changes the behaviour of long messages, since the maximum length of the Message field for opsgenie is 140 characters.
in both of these cases, the Description will be used for the full message, and the message will be set to the title (or to the first 130 characters of the message if no title is provided).